Move out deprecated event structs.
