+++ src/pubd/mod.rs
@@ -7,7 +7,6 @@ mod repository;
 pub use self::commands::{RepoAccessCmd, RepoAccessCmdDet};
 pub use self::events::{
     RepositoryAccessEvent, RepositoryAccessEventDetails, RepositoryAccessIni, RepositoryAccessInitDetails,
-    RrdpSessionReset, RrdpUpdate,
 };
 pub use self::manager::RepositoryManager;
 pub use self::publishers::Publisher;
